% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_shap.R
\name{h2o.shap.dependency_plot}
\alias{h2o.shap.dependency_plot}
\title{shap.dependency.plot}
\usage{
h2o.shap.dependency_plot(shap_values, X, i, log = F)
}
\arguments{
\item{shap_values}{shap value matrix from \code{h2o.predict_contribution}}

\item{X}{original dataset}

\item{i}{variable to be plotted}

\item{log}{whether log transformation is required}
}
\description{
shap dependency plot
}
\examples{
plots <- list()
for(i in c('cyl','disp','hp','wt')){
 log <- ifelse(i == 'disp', TRUE, FALSE)
 plots[[i]] <- h2o.shap.dependency_plot(shap_values_mtcars, mtcars, i, log)
}
ggpubr::ggarrange(plotlist = plots, ncol = 2, nrow = 2)
}
