% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_shap.R
\name{h2o.shap.varimp_plot}
\alias{h2o.shap.varimp_plot}
\title{h2o.shap.varimp_plot}
\usage{
h2o.shap.varimp_plot(shap_values, X, max.display = 20, plot_type = "bar")
}
\arguments{
\item{shap_values}{shap value matrix from \code{h2o.predict_contribution}}

\item{X}{original dataset}

\item{max.display}{number of variables to show in the plot}

\item{plot_type}{{'dot','bar'}}
}
\description{
h2o shap variable importance plot
}
\examples{
data(shap_values_mtcars)
data(mtcars)
h2o.shap.varimp_plot(shap_values_mtcars, mtcars, 20, 'bar')
h2o.shap.varimp_plot(shap_values_mtcars, mtcars, 20, 'dot')
}
