% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_shap.R
\name{h2o.shap.waterfall_plot}
\alias{h2o.shap.waterfall_plot}
\title{h2o.shap.waterfall_plot}
\usage{
h2o.shap.waterfall_plot(
  shap_values,
  X,
  idx,
  link = "",
  centre = FALSE,
  max.display = 20,
  type = "bar"
)
}
\arguments{
\item{shap_values}{shap value matrix from \code{h2o.predict_contribution}}

\item{X}{original dataset}

\item{idx}{index of the individual prediction}

\item{link}{linking function {'', 'logit'}. default NULL: identical, alternative logit: logit transform to probability}

\item{centre}{whether shap.value centred at 0, probabilty centred at 0.5, default FALSE FALSE: add bias term back}

\item{max.display}{number of variables to show in the plot}

\item{type}{{bar, arrow}}
}
\value{
ggplot2 waterfall plot
}
\description{
h2o shap waterfall plot
}
\examples{
data(mtcars)
data(shap_values_mtcars)
h2o.shap.waterfall_plot(shap_values_mtcars, X = mtcars, idx = 1)
}
